requirements.txt
accelerate==1.12.0
aiofiles==25.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
altair==6.0.0
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.12.1
async-timeout==4.0.3
attrs==25.4.0
backoff==2.2.1
beautifulsoup4==4.14.3
blinker==1.9.0
bs4==0.0.2
cachetools==6.2.4
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
coloredlogs==15.0.1
contourpy==1.3.2
cryptography==46.0.3
cycler==0.12.1
dataclasses-json==0.6.7
Deprecated==1.3.1
distro==1.9.0
effdet==0.4.1
emoji==2.15.0
et_xmlfile==2.0.0
exceptiongroup==1.3.1
faiss-cpu==1.13.2
filelock==3.20.3
filetype==1.2.0
flatbuffers==25.12.19
fonttools==4.61.1
frozenlist==1.8.0
fsspec==2026.1.0
gitdb==4.0.12
GitPython==3.1.46
google-api-core==2.29.0
google-auth==2.47.0
google-cloud-vision==3.12.0
googleapis-common-protos==1.72.0
greenlet==3.3.0
grpcio==1.76.0
grpcio-status==1.76.0
h11==0.16.0
hf-xet==1.2.0
html5lib==1.1
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.3
huggingface-hub==0.36.0
humanfriendly==10.0
idna==3.11
Jinja2==3.1.6
jiter==0.12.0
joblib==1.5.3
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
kiwisolver==1.4.9
langchain==1.2.6
langchain-classic==1.0.1
langchain-community==0.4.1
langchain-core==1.2.7
langchain-huggingface==1.2.0
langchain-openai==1.1.7
langchain-text-splitters==1.1.0
langchain-unstructured==1.0.1
langdetect==1.0.9
langgraph==1.0.6
langgraph-checkpoint==4.0.0
langgraph-prebuilt==1.0.6
langgraph-sdk==0.3.3
langsmith==0.6.4
lxml==6.0.2
Markdown==3.10
MarkupSafe==3.0.3
marshmallow==3.26.2
matplotlib==3.10.8
ml_dtypes==0.5.4
mpmath==1.3.0
msoffcrypto-tool==6.0.0
multidict==6.7.0
mypy_extensions==1.1.0
narwhals==2.15.0
networkx==3.4.2
nltk==3.9.2
numpy==2.2.6
nvidia-cublas-cu12==12.8.4.1
nvidia-cuda-cupti-cu12==12.8.90
nvidia-cuda-nvrtc-cu12==12.8.93
nvidia-cuda-runtime-cu12==12.8.90
nvidia-cudnn-cu12==9.10.2.21
nvidia-cufft-cu12==11.3.3.83
nvidia-cufile-cu12==1.13.1.3
nvidia-curand-cu12==10.3.9.90
nvidia-cusolver-cu12==11.7.3.90
nvidia-cusparse-cu12==12.5.8.93
nvidia-cusparselt-cu12==0.7.1
nvidia-nccl-cu12==2.27.5
nvidia-nvjitlink-cu12==12.8.93
nvidia-nvshmem-cu12==3.3.20
nvidia-nvtx-cu12==12.8.90
olefile==0.47
omegaconf==2.3.0
onnx==1.20.1
onnxruntime==1.23.2
openai==2.15.0
opencv-python==4.12.0.88
openpyxl==3.1.5
orjson==3.11.5
ormsgpack==1.12.1
packaging==25.0
pandas==2.3.3
pdf2image==1.17.0
pdfminer.six==20260107
pi_heif==1.1.1
pikepdf==10.2.0
pillow==12.1.0
propcache==0.4.1
proto-plus==1.27.0
protobuf==6.33.4
psutil==7.2.1
pyarrow==22.0.0
pyasn1==0.6.2
pyasn1_modules==0.4.2
pycocotools==2.0.11
pycparser==2.23
pydantic==2.12.5
pydantic-settings==2.12.0
pydantic_core==2.41.5
pydeck==0.9.1
pypandoc==1.16.2
pyparsing==3.3.1
pypdf==6.6.0
pypdfium2==5.3.0
python-dateutil==2.9.0.post0
python-docx==1.2.0
python-dotenv==1.2.1
python-iso639==2025.11.16
python-magic==0.4.27
python-multipart==0.0.21
python-oxmsg==0.0.2
python-pptx==1.0.2
pytz==2025.2
PyYAML==6.0.3
RapidFuzz==3.14.3
referencing==0.37.0
regex==2026.1.15
requests==2.32.5
requests-toolbelt==1.0.0
rpds-py==0.30.0
rsa==4.9.1
safetensors==0.7.0
scikit-learn==1.7.2
scipy==1.15.3
sentence-transformers==5.2.0
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.8.1
SQLAlchemy==2.0.45
streamlit==1.53.0
sympy==1.14.0
tenacity==9.1.2
threadpoolctl==3.6.0
tiktoken==0.12.0
timm==1.0.24
tokenizers==0.22.2
toml==0.10.2
torch==2.9.1
torchvision==0.24.1
tornado==6.5.4
tqdm==4.67.1
transformers==4.57.6
triton==3.5.1
typing-inspect==0.9.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
unstructured==0.18.27
unstructured-client==0.42.8
unstructured.pytesseract==0.3.15
unstructured_inference==1.1.4
urllib3==2.6.3
uuid_utils==0.13.0
watchdog==6.0.0
webencodings==0.5.1
wrapt==2.0.1
xlrd==2.0.2
xlsxwriter==3.2.9
xxhash==3.6.0
yarl==1.22.0
zstandard==0.25.0
